+++ frontEnd/src/components/room/ControllSection.tsx
@@ -1,8 +1,9 @@
 import { MediaObject } from '@/hooks/useMedia';
-import MediaControlButton from '../common/MediaControlButton';
 import settingIcon from '@/assets/setting.svg';
 import SettingModal from './modal/SettingModal';
 import useModal from '@/hooks/useModal';
+import MicControlButton from '../common/MicControlButton';
+import VideoControlButton from '../common/VideoControlButton';
 
 export default function ControllSection({ mediaObject }: { mediaObject: MediaObject }) {
   const { stream } = mediaObject;
@@ -11,17 +12,17 @@ export default function ControllSection({ mediaObject }: { mediaObject: MediaObj
   return (
     <div className="flex justify-between p-2 bg-white border rounded-lg drop-shadow-lg">
       <div className="flex gap-2">
-        <MediaControlButton
+        <MicControlButton
           stream={stream as MediaStream}
-          kind="mic"
           className="w-[3vw] p-[1vw] hover:opacity-50 rounded-[15px]  shadow drop-shadow-2xl"
-          color="black"
+          onColor="black"
+          offColor="white"
         />
-        <MediaControlButton
+        <VideoControlButton
           stream={stream as MediaStream}
-          kind="video"
           className="w-[3vw] p-[1vw] hover:opacity-50 rounded-[15px]  shadow drop-shadow-2xl"
-          color="black"
+          onColor="black"
+          offColor="white"
         />
       </div>
       <button type="button" className="w-[3vw] p-[1vw] hover:opacity-50 rounded-[15px] border-white " onClick={() => show({ mediaObject })}>
